fix: photo compression utils
